Rename DomainExtract to HostnameExtract. Handles ipv4
